+++ pages/auth/sign-in/index.tsx
@@ -1,4 +1,4 @@
-import { getLayout } from '@/layout'
+import { getLayout } from '@/components'
 
 function SignInPage() {
   return <div>Sign In page</div>
