+++ pages/docs/terms-of-service/index.tsx
@@ -1,4 +1,4 @@
-import { getLayout } from '@/layout'
+import { getLayout } from '@/components'
 
 function TermsOfServicePage() {
   return <div>Terms of Service page</div>
