+++ pages/index.tsx
@@ -1,4 +1,4 @@
-import { getLayout } from '@/layout'
+import { getLayout } from '@/components/layout'
 import Head from 'next/head'
 
 export default function Home() {
