+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './header'
+export * from './layout'
 export * from './recaptcha'
 export * from './sidebar'
