+++ src/components/layout/sidebar-layout/SidebarLayout.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren, ReactElement } from 'react'
 
 import { SideBar } from '@/components'
-import { Layout } from '@/layout'
+import { Layout } from '@/components/layout'
 import { NextPage } from 'next'
 
 export const SidebarLayout: NextPage<PropsWithChildren> = ({ children }) => {
